Moved LJDataFrame to its own module
